radiation/requirements.txt
iotile-core
iotile-test
iotile-transport-bled112
pytest

